aider: Add the github_release role to the test-role.yml with the appropriate values for the tflint application.
# Aider chat conversation:

USER: Add the github_release role to the test-role.yml with the appropriate values for the tflint application.

ASSISTANT: To add the `github_release` role to the `test-role.yml` file with appropriate values for the `tflint` application, I will provide the necessary diffs.

Here are the diffs for the `test-role.yml` file:

```diff
--- vm-setup/test-role.yml
+++ vm-setup/test-role.yml
@@ ... @@
   roles:
     #- role: docker
     #- role: nodejs
     - role: git_bug
+    - role: github_release
+      vars:
+        github_repo: "terraform-linters/tflint"
+        github_asset_name: "tflint"
+        github_asset_pattern: "tflint_.*_linux_amd64.zip"
+        github_dest_path: "/usr/local/bin/tflint"
+        github_version_tracking_file: "{{ playbook_dir }}/.tflint_version"
+        github_version_tracking_file_mode: '0644'
```

This change adds the `github_release` role to the `test-role.yml` file and sets the variables to appropriate values for downloading and tracking the `tflint` application.