aider: Fix "Set vars" task in github_release role by using set_fact module for version tracking.
# Aider chat conversation:

USER: I ran this command:

 cd vm-setup && ansible-playbook -K test-role.yml

And got this output:

ERROR! no module/action detected in task.

The error appears to be in '/home/harleypig/projects/dotfiles/vm-setup/roles/github_release/tasks/main.yml': line 9, column 3, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

- name: Set vars
  ^ here

ASSISTANT: The error message indicates that there is a problem with the "Set vars" task in the `main.yml` file of the `github_release` role. It seems that the task is missing a module or action. In Ansible, every task must call a module to perform an action. The "Set vars" task should not be a standalone task; instead, it should set variables for another task or use the `set_fact` module to define variables that will be used later in the playbook.

Here is the diff to correct the issue by using the `set_fact` module to define the variables:

```diff
--- vm-setup/roles/github_release/tasks/main.yml
+++ vm-setup/roles/github_release/tasks/main.yml
@@ ... @@

- name: Set vars
-  vars:
-    github_version_tracking_yaml: "{{ lookup('file', github_version_tracking_file) | from_yaml | default({}) }}"
-    github_current_asset_version: "{{ github_version_tracking_yaml[github_asset_name] | default('0') }}"
+ name: Set facts for version tracking
+ set_fact:
+   github_version_tracking_yaml: "{{ lookup('file', github_version_tracking_file) | from_yaml | default({}) }}"
+   github_current_asset_version: "{{ github_version_tracking_yaml[github_asset_name] | default('0') }}"

 - name: Get the latest release of {{ github_repo }} from GitHub
   uri:
```

This change replaces the incorrect "Set vars" task with a "Set facts for version tracking" task that uses the `set_fact` module to define the `github_version_tracking_yaml` and `github_current_asset_version` variables.