aider: Created a new YAML file with the default filename for the `tflint` package and set the version number to `v0.49.0`.
# Aider chat conversation:

USER: Create a yamlfile using github_release default filename for the tflint package. Use v0.49.0 for the version number.

ASSISTANT: To create a new YAML file with the default filename for the `tflint` package and set the version number to `v0.49.0`, I will show a diff as if creating a new file. The default filename is defined in `vm-setup/roles/github_release/defaults/main.yml` as `.github_release_installed_versions`. Here is the diff for the new file:

```diff
--- /dev/null
+++ vm-setup/.github_release_installed_versions
@@ ... @@
+tflint: v0.49.0
```

This diff indicates the creation of a new file named `.github_release_installed_versions` in the `vm-setup` directory with the content `tflint: v0.49.0`.