aider: Use github_asset_name as a key in the github_version_tracking_yaml dictionary.
# Aider chat conversation:

USER: What is the proper way to use github_asset_name as a key to the `github_version_tracking_yaml` dict?

ASSISTANT: To properly use `github_asset_name` as a key to the `github_version_tracking_yaml` dictionary, you need to ensure that the variable `github_asset_name` is defined and accessible at the point where you're trying to use it in the debug task. Since `github_asset_name` is a role variable, it should be available for use within the tasks of the role.

Here's how you can reference `github_asset_name` as a key in the `github_version_tracking_yaml` dictionary:

```diff
--- vm-setup/roles/github_release/tasks/main.yml
+++ vm-setup/roles/github_release/tasks/main.yml
@@ ... @@
 - name: Convert github version tracking file content to YAML
   set_fact:
     github_version_tracking_yaml: "{{ (github_version_tracking_file_content['content'] | b64decode) | from_yaml }}"

-- name: Dump version information
-  debug:
-    msg: "Current version of tflint: {{ github_version_tracking_yaml['tflint'] }}"
+- name: Dump version information using the github_asset_name variable
+  debug:
+    msg: "Current version of {{ github_asset_name }}: {{ github_version_tracking_yaml[github_asset_name] }}"
```

This change will dynamically use the value of `github_asset_name` to access the corresponding version from the `github_version_tracking_yaml` dictionary. If `github_asset_name` is set to `"tflint"`, it will access `github_version_tracking_yaml['tflint']`.